moved from ad-hoc background render to more organised system; setup for hittable objects

+++ include/icarus/vector.hpp
@@ -9,7 +9,7 @@ namespace ic {
 
 class Vec3 {
  public:
-  using ValueType = float;
+  using ValueType = double;
 
   using PointerType = ValueType*;
   using ConstPointerType = ValueType const*;
